1_Purpose_Built_AI_Services_Demo_Apps/AWS_App_Deployment/lambda/preprocessing/requirements.txt
fastapi==0.104.1
mangum==0.17.0
python-multipart==0.0.6
boto3==1.29.7
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
httpx==0.25.2